extract out weatherhour component
